gcp-cloud-functions/fd-io-api-airflow-logs-manager/requirements.txt
google-cloud-firestore>=1.1.0
google-cloud-storage>=1.16.1
firebase-admin>=2.17.0
requests-toolbelt>=0.9.1
Jinja2>=2.10.1
